requirements.txt
Django==2.2.13
djangorestframework
requests
pytest
